Fix signed union discriminator labels
The generated union discriminator labels for negative 8- and 16-bit
discriminants were incompatible with the (de)serializer because the
former sign extended them, whereas the latter did not. The way this sign
extension was done furthermore caused compile-time warnings for implicit
conversion of signed to unsigned.

It could easily be fixed by also performing the sign extension in
the (de)serializer, but not doing that has a small run-time benefit and
results in slightly simpler code. As these cases didn't work at all
before, it is a reasonable assumption that no-one would benefit from
keeping the generated code unchanged.

+++ src/tools/idlc/src/libidlc/libidlc__descriptor.c
@@ -410,6 +410,7 @@ static idl_retcode_t
 stash_case_label32(
   const idl_pstate_t *pstate, struct instructions *instructions, uint32_t index, const idl_const_expr_t *const_expr)
 {
+  // discrimiant value and case labels are always read as unsigned integer by the deserializer
   int cnt = 0;
   struct instruction inst = { CONSTANT, { .constant = { NULL } } };
   char **strp = &inst.data.constant.value;
@@ -449,7 +450,7 @@ stash_case_label32(
         cnt = idl_asprintf(strp, "%s", literal->value.bln ? "true" : "false");
         break;
       case IDL_INT8:
-        cnt = idl_asprintf(strp, "%" PRId8, literal->value.int8);
+        cnt = idl_asprintf(strp, "%" PRIu8, (uint8_t) literal->value.int8);
         break;
       case IDL_OCTET:
       case IDL_UINT8:
@@ -457,7 +458,7 @@ stash_case_label32(
         break;
       case IDL_SHORT:
       case IDL_INT16:
-        cnt = idl_asprintf(strp, "%" PRId16, literal->value.int16);
+        cnt = idl_asprintf(strp, "%" PRIu16, (uint16_t) literal->value.int16);
         break;
       case IDL_USHORT:
       case IDL_UINT16:
@@ -465,7 +466,7 @@ stash_case_label32(
         break;
       case IDL_LONG:
       case IDL_INT32:
-        cnt = idl_asprintf(strp, "%" PRId32, literal->value.int32);
+        cnt = idl_asprintf(strp, "%" PRIu32, (uint32_t) literal->value.int32);
         break;
       case IDL_ULONG:
       case IDL_UINT32:
